Implement RampRate Usage for Drivebase
The NEOs broke one of the drivetrain chains, so we wanted to implement usage of the OpenLoopRampRate, which takes as a parameter the time in seconds to go from 0 throttle to full throttle.

We might also want to implement ramp rate and velocity control or just full power for climb, using the limit switches to zero at the beginning of the match.