Slower Intake Arm Deadzone, Climb1Holding
Now uses copySign in manualarmmovement to properly calculate the output power when getY is negative. I need a code review on this to make sure I didn't screw up

New constants to be tested for keeping the robot at the same height when lifting itself. This command for holding is run as soon as Hubert lets go of the climbWithJoystick button.